chore(tabbar): split background color converter
